fix some-urls-and-names-of-pages
